changes to use stage

+++ core/src/com/crazyteaparty/teanovelengine/engine/sprite/GUI.java
@@ -4,7 +4,7 @@ import com.crazyteaparty.teanovelengine.engine.Config;
 
 public class GUI extends Sprites{
 
-	public GUI(String texturePathFile, float x, float y, float width, float height) {
+	public GUI(String texturePathFile, int x, int y, int width, int height) {
 	
 		super(Config.PATH_TO_GUI_IMAGE + texturePathFile, x, y, width, height);
